Frequency Bug Fiexed
For “Produce”, "Public Latrine Surfaces", and "Municipal and Piped
Water”, the frequency should be “per week” instead of “per month”. I
changed the create_freqTbl function to modify the labels in the pie
chart. And I changed the calculate_exposure function to differ the
calculation for those three pathways. Basically, we divide the number
by 7 and then multiple by 30 to transfer from weekly frequency to
monthly frequency. For "Municipal and Piped Water”, we want to
calculate how many days they drink municipal water, so that the upper
limit is 30. Please let me know whether the changes works for you and
the code.